Fix onvif discovery
