+++ apps/ex_nvr_web/lib/ex_nvr_web/live/onvif_discovery_live.html.heex
@@ -20,11 +20,11 @@
       </.simple_form>
     </.card>
     <!-- Found devices -->
-    <.card class="w-full md:w-3/5 p-4 overflow-scroll">
+    <.card class="w-full md:w-3/5 p-4">
       <h2 class="text-2xl font-bold mb-4 dark:text-white">Found Devices</h2>
       <p :if={@discovered_devices == []} class="text-l dark:text-white">No devices found</p>
 
-      <ul class="bg-white dark:bg-gray-800 h-200 dark:text-white overflow-scroll">
+      <ul class="bg-white dark:bg-gray-800 h-200 dark:text-white">
         <li
           :for={new_device <- @discovered_devices}
           id={new_device.name}
